requirements.txt
fastapi==0.115.0
motor==3.6.0
pydantic==2.9.2
uvicorn==0.32.0
python-dotenv==1.0.1
